Use decimal places as defined by constants
Also check if SompiPerKaspa is multiple of 10
